// src/api/__generated__/story/model/dramaEditSessionResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { BoundActorCollectionResponse } from './boundActorCollectionResponse';
import type { DramaEditContextEpisodeResponse } from './dramaEditContextEpisodeResponse';

export interface DramaEditSessionResponse {
  uploadSessionId?: number;
  uploadExpireSeconds?: number;
  dramaId?: number;
  userId?: number;
  title?: string;
  description?: string;
  coverUrl?: string;
  onlineAt?: number;
  tagIds?: number[];
  episodes?: DramaEditContextEpisodeResponse[];
  actorCollections?: BoundActorCollectionResponse[];
}